refactor: serializers.py
moved serializers.py to user\api\v1\
